[dotnet] Merge the 'RemoveAttributesStep' into the 'StoreAttributesStep'
Removing (and storing) the attributes earlier make it's possible for them not
to be marked. IOW the attribute type itself (and not just the decorations) can
be removed.

One side effect is that we'll have to manually mark the protocol wrapper type
referenced in the [Protocol] attribute, since the [Protocol] attribute is now
removed.

Inspired by / extracted from #14270.