[src] Propagate the [CCallback] and [BlockCallback] attributes to generated code. (#20225)
The generator needs these attributes in certain cases when a third-party library implements a protocol that subclasses a platform protocol where members of the platform protocol has members with parameters with these attributes.

That's a rather long sentence: in fact there are fewer words in the test verifying this behavior!