[dotnet] Show a better error when using a .NET framework version we don't support yet. (#20255)
If a project tried to use a .NET 9 project (say TargetFramework=net9.0-ios), then
we used to show these rather unhelpful errors:

    error NETSDK1147: The target platform identifier ios was not recognized.

The underlying problem is that we don't support .NET 9 yet, so with this fix we now show:

    error NETSDK1202: The current .NET SDK does not support targeting .NET 9.0. Either target .NET 8.0 or lower, or use a version of the .NET SDK that supports .NET 9.0. Download the .NET SDK from https://aka.ms/dotnet/download

which is much more helpful.

This is accomplished by loading our workload even when using newer .NET
versions we don't support, because the .NET build logic will show the better
error later on.