[release/8.0.1xx-xcode15.1] Only load default items if the workload we loaded matches exactly.
The build will load all the AutoImport.props files from all sdk packs (in our
AutoImport.props file we load the Microsoft.Sdk.DefaultItems.props file). This
file is supposed to include all the default items an sdk pack wants to include
by default, properly conditioned so that multiple versions of the sdk pack
don't end up including the same item more than once.

In the past, conditioning on the .NET version worked, because there would only
be one version of a sdk pack installed per .NET version. However this changed
with multi-targeting, now we can have numerous versions of an sdk pack
installed for every .NET version, which means the condition needs to be
updated, on all versions of the sdk pack.

This is already implemented for the `main` and `release/8.0.1xx-xcode15.0`
branches, but we need it in `release/8.0.1xx-xcode15.1`, because `main`
includes a reference to the sdk pack from this branch.