[tests] Adjust MessageHandlerTest.GHIssue16339 slightly. (#20264)
We've recently had a few test failures like this:

    MonoTests.System.Net.Http.MessageHandlerTest.GHIssue16339
     	[FAIL] GHIssue16339() :   IsSuccessStatusCode #1
       Expected: True
       But was:  False

which doesn't provide much info into what went wrong.

It turns out this particular assert for IsSuccessStatusCode is redundant,
because we assert on the exact StatusCode a bit later, and if that assert
fails, we'll know a bit more.

So just remove this redundant assert for IsSuccessStatusCode.