[AVFoundation] Fix delegate signature for AVAudioSourceNode callbacks.
The signature of the callback provided when creating AVAudioSourceNode instances
was wrong, so the callback would get corrupted data, causing crashes.

Fix the callback to have the right signature (not too difficult), and also make the
existing signature work (a bit more complicated).

Note that this is the second time we've tried to fix the delegate signature for the
AVAudioSourceNode callbacks, but now with tests, so hoping there won't be a third
time.