[dotnet] Compute relative paths using absolute paths.
The 'MSBuild::MakeRelative' path takes the current directory into account.

Given the following example:

    $([MSBuild]::MakeRelative('/Users/rolf/testproject/','Info.plist'))

returns:

1. `/Info.plist` if the current directory is the root directory (`/`).
2. `Info.plist` otherwise.

This is a problem, because we specifically look for "Info.plist" during the
build, and if we don't find it, we may end up with this error:

> A bundle identifier is required. Either add an 'ApplicationId' property in the project file, or add a 'CFBundleIdentifier' entry in the project's Info.plist file.

The fix is to instead of use the relative `Info.plist` path in the call to
`[MSBuild]::MakeRelative`, use the full path instead:

    $([MSBuild]::MakeRelative('/Users/rolf/testproject/','/Users/rolf/testproject/Info.plist'))

and that always yields the expected "Info.plist" result.