[msbuild] Add more diagnostic output to FindItemWithLogicalName.
I've seen multiple cases of the following error:

>  A bundle identifier is required. Either add an 'ApplicationId' property in the project file, or add a 'CFBundleIdentifier' entry in the project's Info.plist file.

In all cases this has been been because the build fails to find Info.plist,
when it's actually there.

One source of this problem was found and fixed with #20374, but that's not the
problem in other scenarios.

So I'm adding a bit more diagnostic output to the FindItemWithLogicalName
task, to try to track down why the task fails to find `Info.plist`.