[msbuild] Use the fully resolved path when computing the relative path for the LogicalName property. Fixes #20330. (#20364)
This fixes an issue where we'd compute a LogicalName property that was outside
of the app bundle, resulting in broken resources in the app bundle (in
particular the app icon would be missing).

So instead of something like this:

    LogicalName: ../../../../../../../../../tmp/testapp/obj/Debug/net8.0-ios/iossimulator-arm64/actool/bundle/Assets.car

we'll now get:

    LogicalName: Assets.car

and the icons will be correctly embedded in the app bundle.

Fixes https://github.com/xamarin/xamarin-macios/issues/20330.