[CoreGraphics] Bind CGEventSetIntegerValueField and CGEventSetDoubleValueField. Fixes #12650. (#20424)
Also expose both getter and setter methods using the enum itself, so we don't
have to make getter and setter properties for each of the enum fields.

This required making CGEventField a public enum, so do that.

And finally document all the new APIs, and some of the old ones.

Fixes https://github.com/xamarin/xamarin-macios/issues/12650.