[msbuild] Fix issue referencing static libraries from compressed xcframeworks.
The code incorrectly assumed that the contents of a compressed xcframework would
always be a framework, but this is incorrect (it can be a dynamic library (.dylib),
static library (.a) or a framework (.framework).

Fix this by:

* Change variable and parameter names to clearly indicate that the output from
  an xcframework is a native library.
* Unify code to handle metadata for native libraries from xcframeworks, so that
  the code isn't duplicated (or even worse, different) between compressed and non-compressed
  xcframeworks.

The latter point fixes a problem where we'd compute the incorrect path to the static
library inside a compressed xcframework.