[msbuild] Fix LogicalName for SceneKitAsset items.
The SceneKitAsset items contains all the resources within a .scnasset directory.
However, we need to pass the .scnasset directory to the native tooling, so we compute
the path to the .scnasset directory, and create a new item for it.

This new item must have the correct LogicalName set, which is accomplished by:

* Not deleting the LogicalName from the original item before using it to compute
  the new item's LogicalName.
* Compute the correct LogicalName for the .scnasset, and actually set it.

This fixes an issue where we'd be unable to detect previously processed SceneKitAsset
items in universal builds, because the files wouldn't located where they should be
on disk.