[CI] Split tests to a post-build pipeline. (#20534)
1Es is problematic and has a lot of issues on how we use the matrix
strategies. We have two options, either wait for 1ES to implement the
needed changes or move all the tests out to two different pipelines that
won't need to be 1ES complient.

In the long run moving out of the build piepline is going to give us
more flexibility and will move the project away from depending on a
thridparty to fix the yaml issues.

Even when landing this PR we will not have the tests removed, we will do
that after we have been able to test that things work as expected.