[CI] Fix mac tests on  post build pipeline
To be able to download a artifact from a build that triggered
the post build pipeline we cannot use the DownloadPipelineArtifact task
with the same values since the artifact wont be found. This change adds
an if statement to do the right thing based on the trigger so that we
can have both pipelines running correctly until we remove the tests
from the build.