[tests] Skip framework-test on iOS and tvOS.
This test doesn't work when building on iossimulator-arm64 or
tvossimulator-arm64, because we're trying to use a fat framework which doesn't
have that architecture (it contains the device arm64 architecture instead).

So just disable this test on iOS and tvOS - the solution is using an
xcframework, and we already have a different test for that.