[CI] Use the cache from devops to try and reduce the amount of times we download mono (#20600)
We do more builds that we bump mono. Moving to a cache will reduce the
amount of network operations we need to do outside our data center. This
should reduce the time it takes to get the files as well as reduce the
number of network errors we get when accessing the mono download.