[maestro] Depend on 'VS.Tools.Net.Core.SDK.Resolver' instead of 'Microsoft.Dotnet.Sdk.Internal'.
dotnet/installer is no longer producing a `Microsoft.Dotnet.Sdk.Internal`
"package" for the SDK, which we're using to get the .NET version to provision
.NET.

So use `VS.Tools.Net.Core.SDK.Resolver` instead, as this is a component
inserted into Visual Studio that contains the same version number.

We may have to change this again in the future, as dotnet/installer is in the
process of merging and/or moving to dotnet/sdk.

Ref:

* https://github.com/xamarin/xamarin-android/pull/8949
* [MS Teams thread](https://teams.microsoft.com/l/message/19:afba3d1545dd45d7b79f34c1821f6055@thread.skype/1715789991637?tenantId=72f988bf-86f1-41af-91ab-2d7cd011db47&groupId=4d73664c-9f2f-450d-82a5-c2f02756606d&parentMessageId=1715789991637&teamName=.NET%20Core%20Eng%20Services%20Partners&channelName=First%20Responders&createdTime=1715789991637)