[tests] Update test to cope with updated linker keeping parameter names.
The linker changed to preserve parameter names for rooted members, so adjust
our test that verifies parameter names are removed to cope.

The method whose parameter names are supposed to be removed can't live inside
the rooted type (i.e. with a [Preserve] attribute), so just put it in a
different class, which isn't rooted/preserved.

Ref: https://github.com/dotnet/runtime/commit/45c608b3633b81fabd24ae1b07fa954cc12e48f9