[CI] Move the API diff outside of the nomarl build CI.
The api diff generates a number of artefacts that are not released. The
way it works is that it compiles the project twice and compares the
reuslts, therefore it has not depency on the build pipeline or in the
tests. Moving the step to its own pipeline simplifies the migration to
the 1ES template AND allows use to identify exactly what failed.