Migrate Messaging Build agent to .NET 8
This is part of the effort to migrate the Pair to Mac agents .NET.

As the Xamarin.iOS.Tasks.Windows project targets netstandard2.0, I'm removing the Build agent reference, and modifying the Makefile to take it from it's output directory. Note: the agent zip file is generated in the intermediate output directory.

# Conflicts:
#	msbuild/Messaging/Xamarin.Messaging.Build/Xamarin.Messaging.Build.csproj