[CI] Move the API diff outside of the normal build CI.  (#20699)
The api diff generates a number of artefacts that are not released. The
way it works is that it compiles the project twice and compares the
reuslts, therefore it has no dependency on the build pipeline or in the
tests. Moving the step to its own pipeline simplifies the migration to
the 1ES template AND allows use to identify exactly what failed.

The two new pipelines are:

- ci:
https://dev.azure.com/devdiv/DevDiv/_build/results?buildId=9690468&view=results
- pr: https://dev.azure.com/devdiv/DevDiv/_build?definitionId=22268

---------

Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>