[msbuild] Add support for signing with the placeholder key.
Add support for signing with a placeholder key ("-") from the developer, and
then use this to sign the prebuilt HotRestart app this way.

The app will have to be signed anyway on the end user machine, so this
shouldn't have any real effect.

This simplifies our build (both locally and on bots), because we won't need a
certificate around to do the actual signing.