Removed VerifyXcodeVersion task and target
This task was executed only for remote builds, adding extra time to the build that is really unnecessary because the SDK is not doing a similar check for local builds. Instead, it fails directly if anything in the current SDK is not compatible.
To maintain the same approach and consistency between local and remote builds, and also to improve remote builds performance, we decided to remove this usage.

Together with the improvements in the build signature computation and SayGoodbye task, the following bug is completed: Bug #1927741 - [XMA] Verify Build Signature + Xcode version check should be faster:https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1927741