[CI] Export more data to the configuration json.
We do not need to perform the configuration on cascade triggered
pipelines, we can add all the needed data in the json and later loaded.
This PR adds all the build variables from the parent build so that we
can later access them in the test pipelines and make the correct
decisions.