[msbuild] Fix computing which items need code signing.
We have logic that tries to detect that if a dylib inside an app needs to be
signed, then the app needs to be signed too.

Unfortunately there was a bug where we'd still take into account dylibs that
had already been signed, which made us always (re)sign any app with a dylib.

So fix this by properly removing items that don't need signing from the list
of items to sign.