[ci] Force CredScan to work. (#20787)
Check this out: the path Azure DevOps checks out a repository to
depends on how many repositories Azure DevOps checks out.

Combine this with the fact that we run CredScan from different jobs [1], and
they don't check out the same number of repositories.

Also combine this with the fact that CredScan needs to access files from
within the repository.

There's probably a nice fix here somewhere, but we're in a hurry, so I figured
out ~a horrible~ an amazing hack: duplicate the files so that they're
available in either location CredScan looks for them.

Seems like this [works](https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=9789287&view=results)!

[1]: it's apparently disabled in the one we're trying to fix, but I have no
idea why that's not working.

---------

Co-authored-by: Alex Soto <alex@soto.dev>