[CI] Add a step in the teardown to take a screenshot
There are ocassions in which the CI fails because there was macOS dialog
that blocks a step. When this happens developers click on it and forget
about it, yet those dialogs can break the CI. this change adds an extra
step that takes a screenshot everytime we have completed a run, that
includes when there is a failure. Later the screenshot can be used to
indentify problems with the bot setup.

Try using pwsh.