[CI] Start the process of cleaning up the config step. (#21134)
We want to be able to export and load the variables in the cacading
pipelines. The way we have the scripts makes this process very
complicated to do. We need to clean up the script. Then export the steps
so that we can access them in the following pipelines.

This change is simply refactoring the way we load/export the default
VSTS variables in a pipeline.

---------

Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>